Add rules for discounts on items and orders

+++ bookstore-backend/src/main/java/com/sbz/bookstore/model/Item.java
@@ -28,7 +28,7 @@ public class Item extends BaseEntity{
 	Order order;
 
 	@Column
-	int discount;
+	double discount;
 
 	@Column
 	int quantity;
